requirements.txt
tensorflow-cpu
tensorflow-hub==0.10.0
Flask==1.1.1
gunicorn==20.0.4
numpy==1.18.0
opencv-python==4.4.0.42
Werkzeug==1.0.1

